refactor: moved gameManager and related methods to separate gameManager directory
