+++ src/lib/gameManager/__test__/getBoardState.test.ts
