+++ src/lib/gameManager/complete.ts
