+++ src/lib/gameManager/create.ts
