+++ src/lib/gameManager/getBoardState.ts
