+++ src/lib/gameManager/observe.ts
