+++ src/lib/socket/getGames.ts
@@ -1,4 +1,4 @@
-import { gameManager } from '@/lib/game/gameManager';
+import { gameManager } from '@/lib/gameManager/gameManager';
 import type { SocketHandler } from '@/types/socket';
 
 export const getGames: SocketHandler['get:games'] = (client) => () => {
